# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Setlists API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/setlist": {
            "post": {
                "tags": [
                    "Setlists"
                ],
                "description": "Creates a setlist with existing songs.",
                "operationId": "App\\Http\\Controllers\\SetlistController::createSetlist",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "A valid Version 4 Uuid",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "The name of the setlist",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "An optional description of the setlist",
                                        "type": "string"
                                    },
                                    "date": {
                                        "description": "A date in format yyyy-mm-dd in wich is supposed to be the show.",
                                        "type": "string"
                                    },
                                    "acts[0][0]": {
                                        "description": "This will add the first song for the first act. You must provide existing song uuids for the acts.",
                                        "type": "string"
                                    },
                                    "acts[0][1]": {
                                        "description": "Same here. In a real use case, any number of acts and songs can be provided.",
                                        "type": "string"
                                    },
                                    "acts[1][0]": {
                                        "description": "The first song of the second act.",
                                        "type": "string"
                                    },
                                    "acts[1][1]": {
                                        "description": "The second song of the second act.",
                                        "type": "string"
                                    },
                                    "acts[1][2]": {
                                        "description": "The third song of the second act. Please note that it's not mandatory to send all songs and acts in the exact order. 'Gaps' in the list are alllowed.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The setlist has been succesfully created.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Error: id or name already exists."
                    },
                    "500": {
                        "description": "Error: invalid identifier, invalid name, invalid id of a song, non existing song, repeated song in setlist."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/setlist/{uuid}": {
            "get": {
                "tags": [
                    "Setlists"
                ],
                "description": "Returns an existing setlist with all its acts and songs.",
                "operationId": "App\\Http\\Controllers\\SetlistController::getSetlist",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the setlist.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A setlist with all its attributes.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: the requested setlist does not exist."
                    },
                    "500": {
                        "description": "Error: invalid identifier."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Setlists"
                ],
                "description": "Deletes an existing setlist. The related songs remains.",
                "operationId": "App\\Http\\Controllers\\SetlistController::deleteSetlist",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the setlist.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The setlist has been succesfully deleted.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: the requested setlist does not exist."
                    },
                    "500": {
                        "description": "Error: invalid identifier."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Setlists"
                ],
                "description": "Updates an existing setlist.",
                "operationId": "App\\Http\\Controllers\\SetlistController::updateSetlist",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the setlist.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the setlist",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "An optional description of the setlist",
                                        "type": "string"
                                    },
                                    "date": {
                                        "description": "A date in format yyyy-mm-dd in wich is supposed to be the show.",
                                        "type": "string"
                                    },
                                    "acts[0][0]": {
                                        "description": "This will add the first song for the first act. You must provide existing song uuids for the acts.",
                                        "type": "string"
                                    },
                                    "acts[0][1]": {
                                        "description": "Same here. In a real use case, any number of acts and songs can be provided.",
                                        "type": "string"
                                    },
                                    "acts[1][0]": {
                                        "description": "The first song of the second act.",
                                        "type": "string"
                                    },
                                    "acts[1][1]": {
                                        "description": "The second song of the second act.",
                                        "type": "string"
                                    },
                                    "acts[1][2]": {
                                        "description": "The third song of the second act. Please note that it's not mandatory to send all songs and acts in the exact order. 'Gaps' in the list are alllowed.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The setlist has been succesfully updated.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Error: name already exists in another stored setlist."
                    },
                    "404": {
                        "description": "Error: the requested setlist does not exist."
                    },
                    "500": {
                        "description": "Error: invalid identifier, invalid name, invalid id of a song, non existing song, repeated song in setlist."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/setlists": {
            "get": {
                "tags": [
                    "Setlists"
                ],
                "description": "Returns all stored setlists (with all its acts and songs) or a range of them, if the optional parameter 'interval' was provided. Filtering by setlist name is also possible.",
                "operationId": "App\\Http\\Controllers\\SetlistController::getSetlists",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interval",
                        "in": "query",
                        "description": "The offset and limit of the requested setlist collection, separated with a comma.",
                        "style": "form",
                        "explode": "false",
                        "allowReserved": "true",
                        "schema": {
                            "type": "string"
                        },
                        "examples": [
                            {
                                "summary": "Getting only the first setlist",
                                "value": "0,1"
                            },
                            {
                                "summary": "A list of setlists, starting in the 11th and with a length of 50 setlists.",
                                "value": "10,50"
                            }
                        ]
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "A string to filter the name of the setlists in the collection. The comparison is case insensitive.",
                        "style": "form",
                        "explode": "false",
                        "allowReserved": "true",
                        "schema": {
                            "type": "string"
                        },
                        "examples": [
                            {
                                "summary": "Getting only those setlists with the string 'example' in their names.",
                                "value": "example"
                            }
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A collection of setlists with all their attributes, according to the 'interval' and 'name' parameters, if provided.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
            "expired": {
                "description": "Error: the provided token has expired, or is invalid or is obsolete."
            },
            "unauthorized": {
                "description": "Error: unauthorized request (no token present)"
            },
...
        }
    },
...
}